+++ tests/testthat/test-bundlePackage.R
@@ -25,12 +25,24 @@ test_that("includes inferred dependencies", {
   expect_true("shiny" %in% names(out))
 })
 
+test_that("recommended packages are snapshotted", {
+  app_dir <- withr::local_tempdir()
+  writeLines(con = file.path(app_dir, "index.Rmd"), c(
+    "```{r}",
+    "library(MASS)",
+    "```"
+  ))
+  out <- bundlePackages(app_dir, appMode = "rmd-static", hasParameters = TRUE)
+  expect_true("MASS" %in% names(out))
+})
+
 test_that("errors if dependencies aren't installed", {
   mockr::local_mock(snapshotRDependencies = function(...) {
     data.frame(
       Package = c("doesntexist1", "doesntexist2"),
       Source = "CRAN",
-      Repository = "https://cran.rstudio.com"
+      Repository = "https://cran.rstudio.com",
+      stringsAsFactors = FALSE
     )
   })
 
@@ -52,8 +64,9 @@ test_that("warns if can't find source", {
   mockr::local_mock(snapshotRDependencies = function(...) {
     data.frame(
       Package = "shiny",
-      Source = "foo",
-      Repository = NA
+      Source = NA,
+      Repository = NA,
+      stringsAsFactors = FALSE
     )
   })
 
@@ -67,6 +80,22 @@ test_that("warns if can't find source", {
   expect_snapshot(
     . <- bundlePackages(app_dir, appMode = "rmd-static")
   )
+})
+
+test_that("clear error if can't run performPackratSnapshot()", {
+  dir <- withr::local_tempdir()
+
+  expect_snapshot(
+    addPackratSnapshot(dir, "doesntexist"),
+    error = TRUE,
+    transform = function(x) gsub('"', "'", x, fixed = TRUE)
+  )
+})
+
+test_that("cleans up implicit dependency files", {
+  dir <- withr::local_tempdir()
+  addPackratSnapshot(dir, "rlang")
+  expect_equal(list.files(dir), "packrat")
 })
 
 test_that("infers correct packages for each source", {